Refactor setup script timeout to 20 minutes for both Ubuntu and macOS (#5)
* Refactor setup script timeout to 20 minutes for both Ubuntu and macOS

* Refactor setup script to enable Gatekeeper

* Refactor setup script to remove Gatekeeper enablement

* Refactor setup script to handle caching and downloading HackGenNerd Font

* Refactor setup script to handle caching and downloading HackGenNerd Font in Linux environment

* Refactor setup script to handle caching and downloading HackGenNerd Font in Linux environment

* Refactor setup script to fix wget command in Linux environment

* Refactor setup script to handle caching and downloading HackGenNerd Font in Linux environment

* Refactor setup script to remove unnecessary --cask flag when installing aws-vault in Linux environment

* Refactor setup script to handle unsupported OS when installing Google Chrome

* Refactor setup script to fix cache key for HackGenNerd Font in Linux environment

* Refactor setup script to cache and download HackGenNerd Font in Linux environment

* Refactor setup script to enable Gatekeeper on macOS

+++ setup.sh
@@ -274,7 +274,25 @@ fi
 echo "aws --version: $(aws --version)"
 
 # aws-vault
-install_if_missing "aws-vault" "brew install aws-vault"
+if ! command -v aws-vault &> /dev/null; then
+    echo "aws-vault がインストールされていません。インストールを試みます。"
+    if [[ "$OS_TYPE" == "Linux" ]]; then
+        brew install aws-vault
+    elif [[ "$OS_TYPE" == "Darwin" ]]; then
+        brew install --cask aws-vault
+    else
+        echo "❌ サポートされていないOSです。aws-vault のインストールをスキップします。"
+    fi
+
+    if ! command -v aws-vault &> /dev/null; then
+        echo "❌ aws-vault のインストールに失敗しました。手動でインストールしてください。"
+        exit 1
+    else
+        echo "✅ aws-vault のインストールが完了しました。"
+    fi
+else
+    echo "✅ aws-vault は既にインストールされています。"
+fi
 
 # jq
 install_if_missing "jq" "brew install jq"
@@ -438,6 +456,8 @@ if ! command -v google-chrome &> /dev/null; then
         else
             echo "✅ Google Chrome のインストールが完了しました。"
         fi
+    else
+        echo "❌ サポートされていないOSです。Google Chrome のインストールをスキップします。"
     fi
 else
     echo "✅ Google Chrome は既にインストールされています。"
@@ -523,15 +543,23 @@ if [[ "$OS_TYPE" == "Linux" ]]; then
         HACKGEN_VERSION="2.9.0"
         DL_PATH="$HOME/Downloads"
 
-        # HackGen_NF のダウンロードとインストール
-        wget -P "$DL_PATH" "https://github.com/yuru7/HackGen/releases/download/v${HACKGEN_VERSION}/HackGen_NF_v${HACKGEN_VERSION}.zip"
+        # CI のときは事前にダウンロードしてキャッシュしてるのでスキップ
+        if [ "$CI" == "true" ]; then
+            echo "CI環境で実行されているため、ダウンロードをスキップします。"
+        else
+            # HackGen_NF のダウンロードとインストール
+            wget -P "$DL_PATH" "https://github.com/yuru7/HackGen/releases/download/v${HACKGEN_VERSION}/HackGen_NF_v${HACKGEN_VERSION}.zip"
+        fi
         unzip -o "${DL_PATH}/HackGen_NF_v${HACKGEN_VERSION}.zip" -d "$DL_PATH"
         # ユーザーにインストール
         mkdir -p "$HOME/.local/share/fonts"
         cp -r "${DL_PATH}/HackGen_NF_v${HACKGEN_VERSION}/"* "$HOME/.local/share/fonts/"
-        # インストーラとディレクトリを削除
-        rm -rf "${DL_PATH}/HackGen_NF_v${HACKGEN_VERSION}"
-        rm -rf "${DL_PATH}/HackGen_v${HACKGEN_VERSION}.zip"
+
+        if [ "$CI" != "true" ]; then
+            # インストーラとディレクトリを削除
+            rm -rf "${DL_PATH}/HackGen_NF_v${HACKGEN_VERSION}"
+            rm -rf "${DL_PATH}/HackGen_v${HACKGEN_VERSION}.zip"
+        fi
 
         # HackGen のダウンロードとインストール
         wget -P "$DL_PATH" "https://github.com/yuru7/HackGen/releases/download/v${HACKGEN_VERSION}/HackGen_v${HACKGEN_VERSION}.zip"
